ingestion/requirements.txt
pdfplumber
sentence-transformers
qdrant-client
langchain-text-splitters
requests
beautifulsoup4
GitPython
